+++ tests/framework/db/mssql/SchemaTest.php
@@ -111,6 +111,7 @@ class SchemaTest extends \yiiunit\framework\db\SchemaTest
     public function getExpectedColumns()
     {
         $columns = parent::getExpectedColumns();
+
         unset($columns['enum_col']);
         unset($columns['ts_default']);
         unset($columns['bit_col']);
@@ -120,17 +121,15 @@ class SchemaTest extends \yiiunit\framework\db\SchemaTest
         $columns['int_col2']['dbType'] = 'int';
         $columns['tinyint_col']['dbType'] = 'tinyint';
         $columns['smallint_col']['dbType'] = 'smallint';
-        $columns['float_col']['dbType'] = 'decimal';
+        $columns['float_col']['dbType'] = 'decimal(4,3)';
         $columns['float_col']['phpType'] = 'string';
         $columns['float_col']['type'] = 'decimal';
-        $columns['float_col']['scale'] = null;
         $columns['float_col2']['dbType'] = 'float';
         $columns['float_col2']['phpType'] = 'double';
         $columns['float_col2']['type'] = 'float';
-        $columns['float_col2']['scale'] = null;
         $columns['blob_col']['dbType'] = 'varbinary';
-        $columns['numeric_col']['dbType'] = 'decimal';
-        $columns['numeric_col']['scale'] = null;
+        $columns['numeric_col']['dbType'] = 'decimal(5,2)';
+        $columns['numeric_col']['scale'] = 2;
         $columns['time']['dbType'] = 'datetime';
         $columns['time']['type'] = 'datetime';
         $columns['bool_col']['dbType'] = 'tinyint';
@@ -141,13 +140,13 @@ class SchemaTest extends \yiiunit\framework\db\SchemaTest
         });
 
         array_walk($columns, static function (&$item, $name) {
-            if (!in_array($name, ['char_col', 'char_col2', 'char_col3'])) {
+            if (!in_array($name, ['char_col', 'char_col2', 'char_col3', 'float_col', 'numeric_col'])) {
                 $item['size'] = null;
             }
         });
 
         array_walk($columns, static function (&$item, $name) {
-            if (!in_array($name, ['char_col', 'char_col2', 'char_col3'])) {
+            if (!in_array($name, ['char_col', 'char_col2', 'char_col3', 'float_col', 'numeric_col'])) {
                 $item['precision'] = null;
             }
         });